+++ src/ui/components/csurf_ui_list_selectable.hpp
@@ -57,13 +57,11 @@ static void ReaSonusListSelectable(
 
     const bool selected = item_index == *selected_item;
     const bool mouse_over = (!popup_open
-                             && between(static_cast<int>(pos_screen_x), width, static_cast<int>(mouse_pos_x))
-                             && between(static_cast<int>(pos_screen_y), selectable_height,
-                                        static_cast<int>(mouse_pos_y))) || (
-                                popup_open && *hovered_item == item_index);
-    const bool context_over = between(static_cast<int>(pos_screen_x), width, static_cast<int>(mouse_pos_x))
-                              && between(static_cast<int>(pos_screen_y), selectable_height,
-                                         static_cast<int>(mouse_pos_y));
+                             && between(pos_screen_x, width, mouse_pos_x)
+                             && between(pos_screen_y, selectable_height, mouse_pos_y)) 
+                             || (popup_open && *hovered_item == item_index);
+    const bool context_over = between(pos_screen_x, width, mouse_pos_x)
+                              && between(pos_screen_y, selectable_height, mouse_pos_y);
 
     /**
      * Definings some colors§
